fix(review): keep projection contracts within package bounds
