+++ src/autoskillit/workspace/_projection_cache.py
@@ -28,30 +28,33 @@ from autoskillit.core import (
     RetiringAppendResult,
     RetiringArtifactRecord,
     _InstallLock,
+    directory_tree_digest,
     get_logger,
-)
-from autoskillit.workspace._projection_artifact import (
-    PROJECTION_ARTIFACT_MANIFEST_SCHEMA_VERSION,
-    projected_artifact_lease_path,
-    projected_artifact_manifest_path,
-    projected_plugin_artifact_digest,
-    read_projected_plugin_identity,
+    is_canonical_plugin_artifact_digest,
+    is_canonical_plugin_artifact_incarnation_id,
+    read_versioned_json,
 )
 
 logger = get_logger(__name__)
 
 __all__ = [
     "PROJECTION_CACHE_KEY_EXCLUSIONS",
+    "PROJECTION_ARTIFACT_MANIFEST_SCHEMA_VERSION",
     "ProjectionCacheKey",
     "ProjectedPluginRetirementOwner",
     "is_projected_asset",
     "iter_public_plugin_asset_files",
+    "projected_artifact_lease_path",
+    "projected_artifact_manifest_path",
+    "projected_plugin_artifact_digest",
     "prune_stale_projections",
     "public_plugin_asset_digest",
+    "read_projected_plugin_identity",
 ]
 
 #: Reserved grace window for lease-aware retirement.
 _PROJECTION_GRACE_HOURS = 6
+PROJECTION_ARTIFACT_MANIFEST_SCHEMA_VERSION = 2
 
 _CANONICAL_SKILL_DIRS = frozenset({"skills", "skills_extended"})
 _PUBLIC_PLUGIN_ASSET_NAMES = frozenset(
@@ -67,6 +70,78 @@ _PUBLIC_PLUGIN_ASSET_NAMES = frozenset(
         "settings.json",
     }
 )
+
+
+def projected_artifact_manifest_path(managed_path: Path) -> Path:
+    """Return the stable sidecar manifest for a projected root."""
+    managed_path = Path(managed_path)
+    return managed_path.parent / f".{managed_path.name}.autoskillit-projection.json"
+
+
+def projected_artifact_lease_path(managed_path: Path) -> Path:
+    """Return the stable lease sidecar for a projected root."""
+    managed_path = Path(managed_path)
+    return managed_path.parent / ".artifact-leases" / f"{managed_path.name}.lock"
+
+
+def projected_plugin_artifact_digest(public_root: Path) -> str:
+    """Hash the complete projection with the canonical artifact-tree contract."""
+    try:
+        return directory_tree_digest(public_root)
+    except (OSError, ValueError) as exc:
+        raise PluginArtifactValidationError(
+            f"projected plugin artifact cannot be digested: {public_root}"
+        ) from exc
+
+
+def read_projected_plugin_identity(
+    managed_path: Path,
+    *,
+    manifest_path: Path,
+    expected_semantic_key: str,
+    expected_projection_version: int | None = None,
+) -> PluginArtifactIdentity:
+    """Read and validate one exact projected artifact identity."""
+    managed_path = Path(managed_path)
+    manifest_path = Path(manifest_path)
+    if manifest_path.is_symlink() or not manifest_path.is_file():
+        raise PluginArtifactValidationError(
+            f"projected plugin identity manifest is not a regular file: {manifest_path}"
+        )
+    manifest = read_versioned_json(
+        manifest_path,
+        PROJECTION_ARTIFACT_MANIFEST_SCHEMA_VERSION,
+    )
+    if manifest is None:
+        raise PluginArtifactValidationError(
+            f"projected plugin identity manifest is unreadable: {manifest_path}"
+        )
+    semantic_key = manifest.get("semantic_key")
+    if semantic_key != expected_semantic_key:
+        raise PluginArtifactValidationError(
+            f"projected plugin semantic key mismatch: {manifest_path}"
+        )
+    incarnation_id = manifest.get("incarnation_id")
+    if not is_canonical_plugin_artifact_incarnation_id(incarnation_id):
+        raise PluginArtifactValidationError(
+            f"projected plugin incarnation is not canonical uuid4 hex: {manifest_path}"
+        )
+    artifact_digest = manifest.get("artifact_digest")
+    if not is_canonical_plugin_artifact_digest(artifact_digest):
+        raise PluginArtifactValidationError(f"projected plugin digest is invalid: {manifest_path}")
+    if (
+        expected_projection_version is not None
+        and manifest.get("projection_version") != expected_projection_version
+    ):
+        raise PluginArtifactValidationError(f"projected plugin version mismatch: {manifest_path}")
+    return PluginArtifactIdentity(
+        semantic_key=semantic_key,
+        incarnation_id=incarnation_id,
+        manifest_schema_version=PROJECTION_ARTIFACT_MANIFEST_SCHEMA_VERSION,
+        artifact_digest=artifact_digest,
+        managed_path=managed_path,
+        manifest_path=manifest_path,
+    )
 
 
 def is_projected_asset(entry: Path, *, top_level: bool) -> bool:
